+++ src/cirrus/plugins/builders/__init__.py
@@ -6,5 +6,5 @@ builder plugins
 
 import cirrus.plugins.builders.venv_pip
 import cirrus.plugins.builders.conda_pip
-import cirrus.plugins.builders.conda_conda
+import cirrus.plugins.builders.conda
 import cirrus.plugins.builders.conda_env
